refactor: header에서 user side 분리(관심사 분리)
